Align LuaEnv/LuaRenamer boundaries; replace LuaContext with LuaSandbox
Every file's namespace now matches its assembly.

LuaEnv owns the whole Lua-side story:
- LuaSerializer moved in from LuaRenamer (its namespace already claimed it),
  reduced to marshaling rules only
- New LuaSandbox : Lua owns the restricted env, Run, CompileFunction
  (memoized by source) and NewTable; its params ctor loads the trusted
  chunks, so LoadChunk is private
- LuaEnv now imports only Shoko *.Enums, not the metadata interface graph
  (its previously unused NLua reference is now load-bearing)

LuaRenamer is host glue:
- ModelProducers moved out of LuaEnv; EnvToModel(args, logger) derives the
  primary series/episode, EpisodeNumbers and the log delegates itself,
  down from 13 positional parameters
- LuaContext deleted; GetPath builds the sandbox and serializer directly
- LuaScripts extracts the lua/ path and trusted-chunk file cache

Tests drop TestSandbox and the InternalsVisibleTo test ctor: LuaSandbox is
a Lua, so retyping the fixture field left every assertion untouched.

Verified: 0 warnings, 79/79 tests, and the post-build generator reproduces
defs.lua/enums.lua/env.lua byte-identically.
