+++ LuaEnv/Attributes/LuaFieldAttribute.cs
@@ -12,7 +12,7 @@ public class LuaFieldAttribute : Attribute
     public string DefaultValue { get; init; } = LuaTypeNames.nil;
 
     /// <summary>
-    /// When true the field is written by the user script (an output) rather than bound by LuaContext.
+    /// When true the field is written by the user script (an output) rather than serialized into the env.
     /// The builder source generator skips generating a setter for these fields.
     /// </summary>
     public bool Output { get; init; }
