+++ LuaRenamer/LuaRenamerSettings.cs
@@ -16,7 +16,7 @@ public class LuaRenamerSettings : IRelocationProviderConfiguration, IConfigurati
 {
     public static LuaRenamerSettings New(IConfigurationService configurationService, IPluginManager pluginManager)
     {
-                var defaultFile = new FileInfo(Path.Combine(LuaContext.LuaPath, "default.lua"));
+                var defaultFile = new FileInfo(Path.Combine(LuaScripts.LuaPath, "default.lua"));
                 if (!defaultFile.Exists) return new();
                 using var text = defaultFile.OpenText();
                 return new() { Script = text.ReadToEnd() };
