+++ Tests/EnvSerializerTests.cs
@@ -20,15 +20,14 @@ namespace LuaRenamer.Tests;
 [TestClass]
 public class EnvSerializerTests
 {
-    private LuaContext _lua = null!;
+    private LuaSandbox _lua = null!;
     private LuaSerializer _serializer = null!;
 
     [TestInitialize]
     public void Init()
     {
-        // The serializer compiles getname against a real sandbox env; the test-only LuaContext ctor stands one
-        // up (lualinq's `from` etc. loaded) without the full Shoko relocation context.
-        _lua = new LuaContext();
+        // The serializer compiles getname against a real sandbox env, so it needs the trusted chunks loaded.
+        _lua = new LuaSandbox(LuaScripts.LuaLinq, LuaScripts.Utils);
         _serializer = new LuaSerializer(_lua);
     }
